Add idapro.vm
Add idapro.vm, that installs IDA Pro 9 using an IDA Pro installer
`ida-pro_9*.exe` (and optionally a license file) in the Desktop.
Make also some changes to idafree.vm to keep both packages consistent.

There are some conflicts between idafree.vm and idapro.vm, as both use
the same tool name, which enable us to have a single shortcut (in the
Tools directory and the taskbar) and right-click menu option. As long as
idapro.vm is installed after idafree.vm (which should be the case using
the installer that uses alphabetical order), the installation is as I
would expect it:
- IDA Pro 9 is used in shortcut in the Tools directory and the taskbar
  (without taskbar duplication)
- The right-click menu option uses IDA launcher to find the highest
  version of IDA available.

I have kept IDA Pro 9 Desktop shortcut as I find it useful to drag
binaries to it.

We need to exclude this new package from testing, as we can't provide an
IDA installer.

+++ packages/idafree.vm/tools/chocolateyinstall.ps1
@@ -2,7 +2,7 @@ $ErrorActionPreference = 'Stop'
 Import-Module vm.common -Force -DisableNameChecking
 
 try {
-  $toolName = 'idafree'
+  $toolName = 'ida'
   $category = 'Disassemblers'
 
   $packageArgs = @{
@@ -21,10 +21,8 @@ try {
   Install-BinFile -Name $toolname -Path $executablePath
 
   # Delete Desktop shortcut
-  $desktopShortcut = Join-Path ${Env:Public} "Desktop\IDA Freeware 8.4.lnk"
-  if (Test-Path $desktopShortcut) {
-    Remove-Item $desktopShortcut -Force -ea 0
-  }
+  $desktopShortcut = Resolve-Path "${Env:Public}\Desktop\IDA Freeware*"
+  if ($null -ne $desktopShortcut) { Remove-Item $desktopShortcut -Force -ea 0 }
 
   # Download ida_launcher.exe to assist with taskbar and right click option and store it in %RAW_TOOLS_DIR%
   # ida_launcher.exe is a custom binary that searches for the latest ida64.exe and executes it
@@ -32,17 +30,14 @@ try {
   $launcherSource = 'https://raw.githubusercontent.com/mandiant/VM-Packages/119ba385de053b01b0d1732d60ad1b1152496dc2/ida_launcher/ida_launcher.exe'
   $launcherPath = Join-Path ${Env:RAW_TOOLS_DIR} "$launcherName.exe"
   $launcherChecksum = "a98241e476150d053d67d149c1b54816c8306db51e0987613ec25a0f8ad22006"
-  Write-Host "[+] Downloading '$launcherSource'"
   Get-ChocolateyWebFile -PackageName $launcherName -FileFullPath $launcherPath -Url $launcherSource -Checksum $launcherChecksum -ChecksumType "sha256"
-
   VM-Assert-Path $launcherPath
 
-  $menuIcon = Join-Path $toolDir "ida.ico" -Resolve
-
-  VM-Install-Shortcut -toolName "ida" -category $category -executablePath $launcherPath -IconLocation $menuIcon
+  $icon = Join-Path $toolDir "$toolName.ico" -Resolve
+  VM-Install-Shortcut -toolName $toolName -category $category -executablePath $launcherPath -IconLocation $icon
 
   # ida64.exe supports both 32 bit and 64 bit in IDA >= 8.2
-  VM-Add-To-Right-Click-Menu $launcherName 'Open with IDA' "`"$launcherPath`" `"%1`"" "$menuIcon"
+  VM-Add-To-Right-Click-Menu $launcherName 'Open with IDA' "`"$launcherPath`" `"%1`"" "$icon"
 } catch {
   VM-Write-Log-Exception $_
 }
